+++ store/src/filters.ts
@@ -3,8 +3,7 @@ import { TFilterHandler, TFilterType } from "./types";
 export const filtersHandlers: { [key: string]: TFilterHandler } = {
 	text: (a: string, b: string) => {
 		if (a) return a.toLowerCase().indexOf(b.toLowerCase()) !== -1;
-
-		return false;
+		return !b;
 	},
 	richselect: (a: string | number, b: string | number) => {
 		if (typeof b !== "number" && !b) return true;